GLOSH scores

Compute Global-Local Outlier Score from Hierarchies. This is based on hierarchical clustering where the minimum cluster size is k. The resulting outlier score is a measure of how anomalous each observation is. The function uses dbscan::hdbscan to do the calculation.

Usage

glosh_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Minimum cluster size. Default: 5.

...

Additional arguments passed to dbscan::hdbscan

Value

Numerical vector containing GLOSH values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
glosh_scores(y)
#>  [1] 0.76165739 0.29885420 0.54373396 0.26449143 0.52651032 0.29953679
#>  [7] 0.22144694 0.48696341 0.07234347 0.00000000 0.60264400 0.20656934
#> [13] 0.57692188 0.14994199 0.00000000 0.63090221 0.24530224 0.33136027
#> [19] 0.39736922 0.86266657 0.30589642 0.05564083 0.63090221 0.49050404
#> [25] 0.53090114 0.65301602 0.63521101 0.45940843 0.38722272 0.92170368
#> [31] 0.40789823 0.40041792 0.42083762 0.26449143 0.38601666 0.57692188
#> [37] 0.16279259 0.06385470 0.26223123 0.71739959 0.00000000 0.59108821
#> [43] 0.57186511 0.87200293 0.06325751 0.06385470 0.59565797 0.00000000
#> [49] 0.56693077 0.95088838