Lookout probabilities

Compute leave-one-out log score probabilities using a Generalized Pareto distribution. These give the probability of each observation being an anomaly.

lookout(
  object = NULL,
  density_scores = NULL,
  loo_scores = density_scores,
  threshold_probability = 0.95
)

Arguments

object

A model object or a numerical data set.

density_scores

Numerical vector of log scores

loo_scores

Optional numerical vector of leave-one-out log scores

threshold_probability

Probability threshold when computing the POT model for the log scores.

Value

A numerical vector containing the lookout probabilities

Details

This function can work with several object types. If object is not NULL, then the object is passed to density_scores to compute density scores (and possibly LOO density scores). Otherwise, the density scores are taken from the density_scores argument, and the LOO density scores are taken from the loo_scores argument. Then the Generalized Pareto distribution is fitted to the scores, to obtain the probability of each observation.

References

Sevvandi Kandanaarachchi & Rob J Hyndman (2022) "Leave-one-out kernel density estimates for outlier detection", J Computational & Graphical Statistics, 31(2), 586-599. http://robjhyndman.com/publications/lookout

Author

Rob J Hyndman

Examples

# Univariate data
tibble(
  y = c(5, rnorm(49)),
  lookout = lookout(y)
)
#> # A tibble: 50 × 2
#>          y lookout
#>      <dbl>   <dbl>
#>  1  5            0
#>  2 -0.141        1
#>  3  0.127        1
#>  4  1.59         1
#>  5  0.337        1
#>  6  0.310        1
#>  7 -0.163        1
#>  8  0.0136       1
#>  9 -0.690        1
#> 10 -0.714        1
#> # ℹ 40 more rows
# Bivariate data with score calculation done outside the function
tibble(
  x = rnorm(50),
  y = c(5, rnorm(49)),
  fscores = density_scores(y),
  loo_fscores = density_scores(y, loo = TRUE),
  lookout = lookout(density_scores = fscores, loo_scores = loo_fscores)
)
#> # A tibble: 50 × 5
#>         x      y fscores loo_fscores lookout
#>     <dbl>  <dbl>   <dbl>       <dbl>   <dbl>
#>  1  0.233  5        4.82        8.37   0    
#>  2 -1.38   0.177    1.32        1.33   1    
#>  3 -0.447  0.605    1.38        1.39   1    
#>  4 -0.576  1.30     1.67        1.69   1    
#>  5  2.54  -0.759    1.46        1.47   1    
#>  6 -1.10  -1.33     1.73        1.76   1    
#>  7 -0.587 -0.598    1.41        1.42   1    
#>  8 -0.538  1.07     1.55        1.56   1    
#>  9 -1.38   2.01     2.23        2.28   0.981
#> 10 -1.07  -0.560    1.40        1.41   1    
#> # ℹ 40 more rows
# Using a regression model
shiraz <- wine_reviews |> filter(variety %in% c("Shiraz", "Syrah"))
fit_wine <- lm(log(price) ~ points, data = shiraz)
shiraz |>
  mutate(lookout_prob = lookout(fit_wine)) |>
  arrange(lookout_prob)
#> # A tibble: 4,496 × 9
#>    country   state         region winery variety points price  year lookout_prob
#>    <chr>     <chr>         <chr>  <chr>  <chr>    <dbl> <dbl> <dbl>        <dbl>
#>  1 Australia South Austra… Eden … Hensc… Shiraz      91   780  2009      0.00566
#>  2 US        California    Centr… Law    Syrah       92   750  2013      0.00693
#>  3 Australia South Austra… Eden … Hensc… Shiraz      96   820  2010      0.0116 
#>  4 Australia South Austra… South… Penfo… Shiraz      98   850  2008      0.0156 
#>  5 Italy     Tuscany       Tosca… Tua R… Syrah       89   300  2011      0.0181 
#>  6 Australia South Austra… South… Penfo… Shiraz      99   850  2010      0.0189 
#>  7 Italy     Tuscany       Tosca… Tua R… Syrah       90   300  2013      0.0221 
#>  8 Italy     Tuscany       Tosca… Tua R… Syrah       90   300  2012      0.0221 
#>  9 France    Rhône Valley  Côte … Domai… Syrah       93   391  2013      0.0247 
#> 10 Australia South Austra… South… Penfo… Shiraz      96   500  2004      0.0287 
#> # ℹ 4,486 more rows