Density score tail probabilities

Compute the probability of a density score at least as extreme as g given a specified distribution. A density score is given by $g = -(f)$ where $f$ is the density or probability mass function of the distribution. The argument distribution specifies which distribution to use. The probabilities are estimate by simulation; n random values are generated from the specified distribution. Consequently, slightly different values will be returned each time unless a random seed is set using [base]{set.seed

Usage

score_tail_prob(g, distribution = NULL, n = 1e+06)

Arguments

g

vector of density scores.

distribution

A distributional object specifying the distribution to use.

n

Number of samples to generate for estimating the probabilities.

Examples

library(distributional)
score_tail_prob(-log(dnorm(1:4)), dist_normal())
#> [1] 0.317562 0.045598 0.002642 0.000071