Surprisal probabilities

Compute the probability of a surprisal at least as extreme as g. A surprisal is given by \(g = -\log(f)\) where \(f\) is the density or probability mass function of the distribution. These probabilities may be computed in three different ways.

Given a specified distribution.

Using a Generalized Pareto Distribution fitted to the most extreme values of g (those above the threshold_probability quantile). This option is used if GPD = TRUE. For g values with tail probability greater than threshold_probability, the value of threshold_probability is returned.

Empirically as the proportion of values above g. This option is used when GPD = FALSE and distribution = NULL.

Usage

surprisal_prob(
  g,
  distribution = NULL,
  GPD = FALSE,
  smallest_prob = 1e-06,
  gridsize = 100001,
  threshold_probability = 0.9
)

Arguments

g

vector of surprisals.

distribution

A distributional object specifying the probability distribution to use.

GPD

Logical value specifying if a Generalized Pareto distribution should be used to estimate the tail probabilities.

smallest_prob

Smallest detectable tail probability to be used if empirical tail proportions are used. If the surprisal has lower probability than this, smallest_prob is returned.

gridsize

Size of grid used in estimating the empirical tail probabilities. A larger number gives more accurate estimates but takes more time.

threshold_probability

Probability threshold when computing the GPD distribution for the surprisals.

Examples

surprisal_prob(-dnorm(1:3, log = TRUE), dist_normal())
#> [1] 0.317310508 0.045500264 0.002699796
tibble(
  y = n01$v1,
  g = -dnorm(y, log = TRUE),
  prob1 = surprisal_prob(g, dist_normal()),
  prob2 = surprisal_prob(g, GPD = TRUE),
  prob3 = surprisal_prob(g)
) |>
  filter(prob1 < 0.01 | prob2 < 0.01 | prob3 < 0.01)
#> # A tibble: 10 × 5
#>        y     g    prob1    prob2   prob3
#>    <dbl> <dbl>    <dbl>    <dbl>   <dbl>
#>  1 -2.89  5.09 0.00387  0.00461  0.00600
#>  2  2.65  4.43 0.00807  0.00991  0.00800
#>  3 -2.59  4.28 0.00953  0.0118   0.0100 
#>  4 -3.01  5.44 0.00263  0.00308  0.00300
#>  5  3.81  8.18 0.000139 0.000135 0.00100
#>  6  2.68  4.50 0.00746  0.00913  0.00700
#>  7 -2.94  5.24 0.00328  0.00389  0.00500
#>  8 -2.60  4.29 0.00943  0.0116   0.00900
#>  9  3.06  5.59 0.00225  0.00261  0.00200
#> 10 -3.00  5.41 0.00273  0.00320  0.00400