Surprisals computed from a data set

Compute surprisals from a data set given a probability distribution. The surprisals are defined as minus the log of the density at each observation.

Usage

# Default S3 method
surprisals(
  object,
  distribution = dist_kde(object, multiplier = 2, ...),
  loo = FALSE,
  ...
)

Arguments

object

A numerical data set, either a vector, matrix or data frame.

distribution

A probability distribution stored as a distributional object.

loo

Should leave-one-out surprisals be computed?

...

Other arguments are passed to dist_kde.

Value

A numerical vector containing the surprisals.

Details

If no distribution is provided, a kernel density estimate is computed. The leave-one-out surprisals (or LOO surprisals) are obtained by estimating the kernel density estimate using all other observations.

See also

Author

Rob J Hyndman

Examples

# surprisals computed from bivariate data set
oldfaithful |>
  filter(duration < 7000, waiting < 7000) |>
  mutate(
    loo_fscores = surprisals(cbind(duration, waiting), loo = TRUE)
  )
#> Error in mutate(filter(oldfaithful, duration < 7000, waiting < 7000),     loo_fscores = surprisals(cbind(duration, waiting), loo = TRUE)): ℹ In argument: `loo_fscores = surprisals(cbind(duration, waiting), loo =
#>   TRUE)`.
#> Caused by error in `-density(distribution, at = object, log = TRUE)`:
#> ! invalid argument to unary operator