weird: Functions and Data Sets for "That's Weird: Anomaly Detection Using R" by Rob J Hyndman

All functions and data sets required for the examples in the book Hyndman (2024) "That's Weird: Anomaly Detection Using R" https://OTexts.com/weird/. All packages needed to run the examples are also loaded.

See also

Author

Maintainer: Rob Hyndman Rob.Hyndman@monash.edu (ORCID) [copyright holder]

Other contributors:

RStudio [copyright holder]