GLOSH scores

Compute Global-Local Outlier Score from Hierarchies. This is based on hierarchical clustering where the minimum cluster size is k. The resulting outlier score is a measure of how anomalous each observation is. The function uses dbscan::hdbscan to do the calculation.

Usage

glosh_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Minimum cluster size. Default: 5.

...

Additional arguments passed to dbscan::hdbscan

Value

Numerical vector containing GLOSH values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
glosh_scores(y)
#>  [1] 0.237106163 0.728691762 0.788871464 0.841915940 0.131277043 0.127403503
#>  [7] 0.231590928 0.393704982 0.528654271 0.929415004 0.790329473 0.905610803
#> [13] 0.241116794 0.220441965 0.483096259 0.709615371 0.900123969 0.000000000
#> [19] 0.746746720 0.002550038 0.890995770 0.241116794 0.000000000 0.752359985
#> [25] 0.747565590 0.940897081 0.406092256 0.761822783 0.298689078 0.659400878
#> [31] 0.793877007 0.661656067 0.711511611 0.848552374 0.223158014 0.000000000
#> [37] 0.436823757 0.071756101 0.141207607 0.298689078 0.135116081 0.000000000
#> [43] 0.844721470 0.816014762 0.331075299 0.814549794 0.890849262 0.766633240
#> [49] 0.706431871 0.969542539