Table of Highest Density Regions

Compute the highest density regions (HDR) for a kernel density estimate. The HDRs are returned as a tibble with one row per interval and columns: prob (giving the probability coverage), density (the value of the density at the boundary of the HDR), For one dimensional density functions, the tibble also has columns lower (the lower ends of the intervals), upper (the upper ends of the interval), mode (the point at which the density is maximized within each interval).

Usage

hdr_table(
  y = NULL,
  density = NULL,
  prob = c(0.5, 0.99),
  h = kde_bandwidth(y, method = "double"),
  H = kde_bandwidth(y, method = "double"),
  ...
)

Arguments

y

Numerical vector or matrix of data

density

Probability density function, either estimated by ks::kde() or a data frame or matrix with numerical columns that can be passed to as_kde().

prob

Probability of the HDR

h

Bandwidth for univariate kernel density estimate. Default is kde_bandwidth.

H

Bandwidth for multivariate kernel density estimate. Default is kde_bandwidth.

...

If y is supplied, other arguments are passed to kde. Otherwise, additional arguments are passed to as_kde.

Value

A tibble

References

Hyndman, R J. (1996) Computing and Graphing Highest Density Regions, The American Statistician, 50(2), 120–126.

Author

Rob J Hyndman

Examples

# Univariate HDRs
y <- c(rnorm(100), rnorm(100, 3, 1))
hdr_table(y = y)
#> # A tibble: 2 × 5
#>    prob  lower upper  mode density
#>   <dbl>  <dbl> <dbl> <dbl>   <dbl>
#> 1  0.99 -2.20   5.21  1.23  0.0549
#> 2  0.5  -0.171  2.79  1.23  0.137 
hdr_table(density = ks::kde(y))
#> # A tibble: 3 × 5
#>    prob  lower upper   mode density
#>   <dbl>  <dbl> <dbl>  <dbl>   <dbl>
#> 1  0.99 -1.77  5.31  -0.150  0.0320
#> 2  0.5  -0.743 0.691 -0.150  0.167 
#> 3  0.5   2.36  3.18   2.82   0.167 
x <- seq(-4, 4, by = 0.01)
hdr_table(density = data.frame(y = x, density = dnorm(x)), prob = 0.95)
#> # A tibble: 1 × 5
#>    prob lower upper  mode density
#>   <dbl> <dbl> <dbl> <dbl>   <dbl>
#> 1  0.95 -1.96  1.96     0  0.0589
# Bivariate HDRs
y <- cbind(rnorm(100), rnorm(100))
hdr_table(y = y)
#> # A tibble: 2 × 2
#>    prob density
#>   <dbl>   <dbl>
#> 1  0.99 0.00925
#> 2  0.5  0.0618 
grid <- seq(-4, 4, by = 0.1)
density <- expand.grid(grid, grid) |>
  mutate(density = dnorm(Var1) * dnorm(Var2))
hdr_table(density = density)
#> # A tibble: 2 × 2
#>    prob density
#>   <dbl>   <dbl>
#> 1  0.99 0.00157
#> 2  0.5  0.0792