Local outlier factors

Compute local outlier factors using k nearest neighbours. A local outlier factor is a measure of how anomalous each observation is based on the density of neighbouring points. The function uses dbscan::lof to do the calculation.

Usage

lof_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Number of neighbours to include. Default: 5.

...

Additional arguments passed to dbscan::lof

Value

Numerical vector containing LOF values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
lof_scores(y)
#>  [1] 1.8602505 1.0836614 0.9738792 0.9835853 1.5532917 2.2764586 0.9431383
#>  [8] 1.2310532 0.9814542 1.5330695 1.1539843 0.9797052 0.9745644 1.0297884
#> [15] 1.0835096 0.9897310 1.0010920 1.0362240 0.9919557 0.9703356 1.1979561
#> [22] 1.6707920 0.9839592 1.6431551 1.1539843 1.0125753 0.9807421 1.2873032
#> [29] 0.9859738 1.6599105 0.9918026 1.1299040 1.1359317 0.9687939 0.9743172
#> [36] 0.9785081 1.6080700 2.1946791 2.3445178 0.9855117 1.6378290 1.5547381
#> [43] 1.3105915 0.9738792 3.6987105 1.0273728 0.9913223 1.0614502 0.9744490
#> [50] 6.2630315