Local outlier factors

Compute local outlier factors using k nearest neighbours. A local outlier factor is a measure of how anomalous each observation is based on the density of neighbouring points. The function uses dbscan::lof to do the calculation.

Usage

lof_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Number of neighbours to include. Default: 5.

...

Additional arguments passed to dbscan::lof

Value

Numerical vector containing LOF values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
lof_scores(y)
#>  [1] 1.0422776 1.0034834 1.0562048 0.9818967 2.8436906 1.7406616 0.9806163
#>  [8] 0.9848488 0.9806163 0.9813153 1.3461122 1.0415181 1.1533184 1.1288753
#> [15] 0.9798321 2.6597058 1.5312729 0.9497792 1.0375096 1.0653497 1.0434496
#> [22] 0.9926572 3.1811246 2.1100453 0.9131549 1.0841007 1.1019065 1.0638561
#> [29] 1.0823773 1.0249598 1.0317759 1.1796167 1.1205482 0.9529312 0.9582916
#> [36] 1.0219737 0.9551666 1.0249598 0.9922363 0.9762812 1.0434496 0.9806315
#> [43] 1.3937964 1.0265903 4.2565452 1.0972672 1.9511576 1.0382044 2.4362761
#> [50] 8.6019599