Local outlier factors

Compute local outlier factors using k nearest neighbours. A local outlier factor is a measure of how anomalous each observation is based on the density of neighbouring points. The function uses dbscan::lof to do the calculation.

lof_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Number of neighbours to include. Default: 5.

...

Additional arguments passed to dbscan::lof

Value

Numerical vector containing LOF values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
lof_scores(y)
#>  [1] 1.0960398 0.9537760 1.1364141 1.0323738 1.3205502 1.3142822 0.9942209
#>  [8] 1.0570333 1.3633330 0.9564843 1.0103909 1.0064789 1.0283268 1.0376479
#> [15] 1.0645657 1.6359312 1.0792635 1.0739328 1.0561731 1.0274082 1.0854243
#> [22] 1.2418543 1.3500016 1.0284119 2.1496699 1.0308315 0.9503692 0.9496635
#> [29] 1.0467195 1.0247414 1.3305170 1.1621216 1.0867718 1.6431271 1.0178906
#> [36] 1.0192961 1.0950086 1.3212366 1.0444006 1.0674662 2.2074120 1.0838283
#> [43] 1.1200204 1.0591617 1.0546690 1.3635097 0.9878530 1.3667847 1.0409272
#> [50] 7.0578461