Local outlier factors

Compute local outlier factors using k nearest neighbours. A local outlier factor is a measure of how anomalous each observation is based on the density of neighbouring points. The function uses dbscan::lof to do the calculation.

Usage

lof_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Number of neighbours to include. Default: 5.

...

Additional arguments passed to dbscan::lof

Value

Numerical vector containing LOF values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
lof_scores(y)
#>  [1] 0.9850388 0.9986150 1.0027496 1.0351862 1.1128197 1.0027496 1.1453585
#>  [8] 0.9993706 3.7363753 0.9612888 3.9091122 0.9815759 1.0524771 1.1133421
#> [15] 2.8070150 0.9741427 0.9568865 1.0216340 0.9881846 1.1260135 1.1050061
#> [22] 1.0314628 0.9987652 1.0242743 1.0333405 1.0212220 1.0307485 0.9568865
#> [29] 0.9654484 1.3044002 1.1467885 1.2035228 0.9858490 1.0021541 1.4002791
#> [36] 1.0078623 0.9969242 0.9796352 1.0414111 1.1517626 1.0044342 1.3990619
#> [43] 1.1496611 0.9850388 0.9632962 0.9979288 1.4450605 1.2464840 1.6578736
#> [50] 6.9239879