Local outlier factors

Compute local outlier factors using k nearest neighbours. A local outlier factor is a measure of how anomalous each observation is based on the density of neighbouring points. The function uses dbscan::lof to do the calculation.

Usage

lof_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Number of neighbours to include. Default: 5.

...

Additional arguments passed to dbscan::lof

Value

Numerical vector containing LOF values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
lof_scores(y)
#>  [1] 1.1731763 1.0038386 1.0344317 0.9940660 1.1692148 1.0065521 1.0032003
#>  [8] 1.3608241 1.0748054 1.0244770 0.9463716 1.3624905 1.0284496 1.0284496
#> [15] 0.9684426 0.9531102 1.3229512 1.7731860 0.9418055 1.0832003 1.0284496
#> [22] 1.5036480 0.9949457 0.9494549 1.4127066 1.1229598 1.0629572 1.4049935
#> [29] 2.0308353 0.9525437 0.9564544 0.9584486 1.2867697 1.0345743 1.0884150
#> [36] 1.3862706 1.3776025 1.0246578 0.9907164 1.1711188 1.0482850 1.1151756
#> [43] 0.9500504 0.9515915 1.0822548 1.3862781 1.0244770 1.0347052 0.9526818
#> [50] 4.1993988