Local outlier factors

Compute local outlier factors using k nearest neighbours. A local outlier factor is a measure of how anomalous each observation is based on the density of neighbouring points. The function uses dbscan::lof to do the calculation.

Usage

lof_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Number of neighbours to include. Default: 5.

...

Additional arguments passed to dbscan::lof

Value

Numerical vector containing LOF values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
lof_scores(y)
#>  [1] 0.9572075 1.3364746 1.6386387 1.0765727 0.9965913 1.0972453 1.0686246
#>  [8] 1.2503049 1.0790212 1.0615800 1.0572546 0.9719775 1.0347321 0.9871951
#> [15] 0.9840835 0.9702183 0.9503905 1.0791654 1.0095755 0.9894366 1.0604157
#> [22] 1.0758223 0.9686135 1.0341678 0.9779004 1.3318098 1.2228967 1.3367942
#> [29] 1.0527295 0.9686135 0.9999369 1.0696613 1.8065307 1.0580371 0.9614098
#> [36] 1.0327777 1.0426200 0.9779804 1.0033623 1.6101772 1.0065486 1.0659307
#> [43] 1.1844121 1.1025775 1.6927330 0.9766522 1.0693017 0.9884719 0.9572075
#> [50] 7.9085878