Local outlier factors

Compute local outlier factors using k nearest neighbours. A local outlier factor is a measure of how anomalous each observation is based on the density of neighbouring points. The function uses dbscan::lof to do the calculation.

Usage

lof_scores(y, k = 10, ...)

Arguments

y

Numerical matrix or vector of data

k

Number of neighbours to include. Default: 5.

...

Additional arguments passed to dbscan::lof

Value

Numerical vector containing LOF values

See also

Author

Rob J Hyndman

Examples

y <- c(rnorm(49), 5)
lof_scores(y)
#>  [1] 1.0406669 0.9900373 1.0117207 1.0766228 0.9536460 2.2297450 1.0362703
#>  [8] 0.9299435 1.0264040 1.0380001 1.5644976 1.9123641 1.1764795 0.9172575
#> [15] 3.4947121 1.0484774 1.2941360 1.0477277 1.2207751 1.0002774 1.2016687
#> [22] 1.0273918 1.0247039 1.0578666 0.9941317 0.9947802 1.1246896 0.9962224
#> [29] 1.0277018 0.9823940 0.9602361 1.0093079 2.4805006 2.4221470 0.9603850
#> [36] 0.9691792 0.9975073 1.0044611 0.9195883 1.1647098 1.0180839 3.3331210
#> [43] 0.9697781 0.9493904 1.2305515 1.0181051 0.9810810 0.9752703 0.9794090
#> [50] 7.8224327