Lookout probabilities

Compute leave-one-out log score probabilities using a Generalized Pareto distribution. These give the probability of each observation being an anomaly.

Usage

lookout(
  object = NULL,
  density_scores = NULL,
  loo_scores = density_scores,
  threshold_probability = 0.95
)

Arguments

object

A model object or a numerical data set.

density_scores

Numerical vector of log scores

loo_scores

Optional numerical vector of leave-one-out log scores

threshold_probability

Probability threshold when computing the POT model for the log scores.

Value

A numerical vector containing the lookout probabilities

Details

This function can work with several object types. If object is not NULL, then the object is passed to density_scores to compute density scores (and possibly LOO density scores). Otherwise, the density scores are taken from the density_scores argument, and the LOO density scores are taken from the loo_scores argument. Then the Generalized Pareto distribution is fitted to the scores, to obtain the probability of each observation.

References

Sevvandi Kandanaarachchi & Rob J Hyndman (2022) "Leave-one-out kernel density estimates for outlier detection", J Computational & Graphical Statistics, 31(2), 586-599. https://robjhyndman.com/publications/lookout/

Author

Rob J Hyndman

Examples

# Univariate data
tibble(
  y = c(5, rnorm(49)),
  lookout = lookout(y)
)
#> # A tibble: 50 × 2
#>           y lookout
#>       <dbl>   <dbl>
#>  1  5         0    
#>  2  0.720     1    
#>  3 -0.796     1    
#>  4 -0.551     1    
#>  5  1.69      0.965
#>  6  1.38      1    
#>  7 -0.786     1    
#>  8 -0.209     1    
#>  9 -0.00835   1    
#> 10  0.520     1    
#> # ℹ 40 more rows
# Bivariate data with score calculation done outside the function
tibble(
  x = rnorm(50),
  y = c(5, rnorm(49)),
  fscores = density_scores(y),
  loo_fscores = density_scores(y, loo = TRUE),
  lookout = lookout(density_scores = fscores, loo_scores = loo_fscores)
)
#> # A tibble: 50 × 5
#>         x      y fscores loo_fscores lookout
#>     <dbl>  <dbl>   <dbl>       <dbl>   <dbl>
#>  1  0.364  5        4.83        6.32   0    
#>  2  1.07   0.280    1.51        1.52   1    
#>  3 -0.457  0.310    1.52        1.52   1    
#>  4 -0.457  0.472    1.54        1.55   1    
#>  5  0.921  2.16     2.30        2.35   0.991
#>  6 -0.487 -1.15     1.69        1.70   1    
#>  7  0.529  1.21     1.74        1.76   1    
#>  8 -1.75  -1.03     1.65        1.66   1    
#>  9 -0.362  0.326    1.52        1.53   1    
#> 10  0.165  0.764    1.60        1.61   1    
#> # ℹ 40 more rows
# Using a regression model
shiraz <- wine_reviews |> filter(variety %in% c("Shiraz", "Syrah"))
#> Error in eval(expr, envir, enclos): object 'wine_reviews' not found
fit_wine <- lm(log(price) ~ points, data = shiraz)
#> Error in eval(mf, parent.frame()): object 'shiraz' not found
shiraz |>
  mutate(lookout_prob = lookout(fit_wine)) |>
  arrange(lookout_prob)
#> Error in eval(expr, envir, enclos): object 'shiraz' not found