Surprisal probabilities

Compute the probability of a surprisal at least as extreme as those observed. A surprisal is given by \(-\log(f)\) where \(f\) is the density or probability mass function of the distribution. The surprisal values are computed from the distribution provided. If no distribution is provided, a kernel density estimate is used.

Usage

surprisal_prob(
  object,
  distribution = NULL,
  loo = FALSE,
  GPD = FALSE,
  threshold_probability = 0.1,
  ...
)

Arguments

object

A model or numerical data set.

distribution

A probability distribution stored as a distributional object. Ignored if object is a model.

loo

Logical value specifying if leave-one-out surprisals should be computed.

GPD

Logical value specifying if a Generalized Pareto distribution should be used to estimate the probabilities.

threshold_probability

Probability threshold when computing the GPD distribution for the surprisals.

...

Other arguments are passed to surprisals.

Details

The surprisal probabilities may be computed in three different ways.

Given the same distribution that was used to compute the surprisal values. Under this option, surprisal probabilities are equal to 1 minus the coverage probability of the largest HDR that contains each value. Surprisal probabilities smaller than 1e-6 are returned as 1e-6.

Using a Generalized Pareto Distribution fitted to the most extreme surprisal values (those with probability less than threshold_probability). This option is used if GPD = TRUE. For surprisal values with probability less than threshold_probability, the value of threshold_probability is returned. Under this option, the distribution is used for computing the surprisal values but not for determining their probabilities. Due to extreme value theory, the resulting probabilities should be relatively insensitive to the distribution used in computing the surprisal values.

Empirically as the proportion of observations with greater surprisal values. This option is used when GPD = FALSE and no distribution is explicitly provided. This is also insensitive to the distribution used in computing the surprisal values.

Examples

# Univariate data
tibble(
  y = c(5, rnorm(49)),
  p = surprisal_prob(y)
)
#> Error in UseMethod("generate"): no applicable method for 'generate' applied to an object of class "NULL"
tibble(
  y = n01$v1,
  prob1 = surprisal_prob(y),
  prob2 = surprisal_prob(y, GPD = TRUE),
  prob3 = surprisal_prob(y, dist_normal()),
  prob4 = surprisal_prob(y, dist_normal(), GPD = TRUE)
) |>
  arrange(prob1)
#> Error in UseMethod("generate"): no applicable method for 'generate' applied to an object of class "NULL"
# Bivariate data
tibble(
  x = rnorm(50),
  y = c(5, rnorm(49)),
  lookout = lookout_prob(cbind(x, y))
)
#> # A tibble: 50 × 3
#>          x       y lookout
#>      <dbl>   <dbl>   <dbl>
#>  1 -0.0185  5       0.0384
#>  2 -0.375  -0.177   1     
#>  3  0.853  -0.859   1     
#>  4 -0.393   0.471   1     
#>  5  0.817   0.746   1     
#>  6 -1.24    1.94    0.465 
#>  7 -0.464  -0.0194  1     
#>  8 -1.00   -0.390   1     
#>  9 -0.926  -1.82    1     
#> 10 -0.0253  0.247   1     
#> # ℹ 40 more rows
# Using a regression model
of <- oldfaithful |> filter(duration < 7200, waiting < 7200)
fit_of <- lm(waiting ~ duration, data = of)
of |>
  mutate(p = surprisal_prob(fit_of)) |>
  arrange(p)
#> Error in mutate(of, p = surprisal_prob(fit_of)): ℹ In argument: `p = surprisal_prob(fit_of)`.
#> Caused by error in `UseMethod()`:
#> ! no applicable method for 'generate' applied to an object of class "NULL"