Surprisals computed from a data set

Compute surprisals from a data set given a probability distribution. The surprisals are defined as minus the log of the density at each observation.

Usage

# Default S3 method
surprisals(
  object,
  distribution = dist_kde(object, multiplier = 2, ...),
  loo = FALSE,
  ...
)

Arguments

object

A numerical data set, either a vector, matrix or data frame.

distribution

A probability distribution stored as a distributional object.

loo

Should leave-one-out surprisals be computed?

...

Other arguments are passed to dist_kde.

Value

A numerical vector containing the surprisals.

Details

If no distribution is provided, a kernel density estimate is computed. The leave-one-out surprisals (or LOO surprisals) are obtained by estimating the kernel density estimate using all other observations.

See also

Author

Rob J Hyndman

Examples

# surprisals computed from bivariate data set
oldfaithful |>
  filter(duration < 7000, waiting < 7000) |>
  mutate(
    loo_fscores = surprisals(cbind(duration, waiting), loo = TRUE)
  )
#> # A tibble: 2,189 × 4
#>    time                duration waiting loo_fscores
#>    <dttm>                 <dbl>   <dbl>       <dbl>
#>  1 2015-01-02 14:53:00      271    5040        14.2
#>  2 2015-01-09 23:55:00      247    6060        11.5
#>  3 2015-02-07 00:49:00      203    5460        12.8
#>  4 2015-02-14 01:09:00      195    5221        13.3
#>  5 2015-02-21 01:12:00      210    5401        12.3
#>  6 2015-02-28 01:11:00      185    5520        13.7
#>  7 2015-03-07 00:50:00      160    5281        15.3
#>  8 2015-03-13 21:57:00      226    6000        11.9
#>  9 2015-03-13 23:37:00      190    5341        13.5
#> 10 2015-03-20 22:26:00      102    3961        13.5
#> # ℹ 2,179 more rows